tfmini : fix scheduling and modernize output
This reduces the scheduling interval to 9ms, such that the driver is always ready to read new data. Running it at exactly 100Hz is not correct since the driver and sensor measurement intervals are not "in sync", causing the driver to miss data. This causes a fill-up of the UART buffer.


Former-commit-id: 3afa0189547695343feb567e1ed636aa99c165ca