simulator_mavlink: take care of battery percentage rounding errors
In 5bb9babc206068e9ee67e1489433be94482a402f I made MAVLink send a
rounded up integer instead of rounded down. This makes sense in practice
because the low battery reactions happen exactly when the reported number
switches in the UI. But here we want to provoke an exact 50% in the UI so
we stop counting at 49.9%, it get's rounded up and we see the expected
result.


Former-commit-id: 1e14b10dee8cae32818f74297dfb0b78086d4c21