ll40ls: Fix rotation argument parse bug
Before (introduced in b312d5a3baa [formerly 7b16c3482d8]), there was no colon after
the `R` argument in the options specification string (ab:R).

The R should be followed by a colon, because in indicates that
the R option requires an argument, which it does.

So I added a colon.


Former-commit-id: 214e9c8244bfca72a319694ae222727f658ca6c1