mc_pos_control: fix derivative spike when regaining velocity estimate
When having no velocity estimate the derivative was updated with zero.
When losing the velocity estimate this is fine since the resulting
derivative spike doesn't get used and acceleration is set to NAN.
But when regaining the velocity estimate the spike from zero to
the first estimated velocity gets used as acceleration in the position
controller and results in a twitch.

To solve this I use the derivative reset I introduced in pr #13522
commit 4ea17b6fcb9cff51aa2786b2c1b761de555303f6 [formerly b64abf48b211b66fbfc7a3b900adbbbdf0e2e14c]


Former-commit-id: 1e06f6bbd233eaee3211ee0609e5a0875e2e7326