HoverThrustEstimator: add a new single state estimator with measurement noise auto-tuning
The purpose of this estimator is to improve land detection and vertical
velocity feedforward

Recovery strategy:
This is required when the setpoint suddenly changes in air or that the
EKF is diverging. A lowpassed test ratio is used as a trigger for the recovery logic
Also, a lowpassed residual is used to estimate the steady-state value
and remove it when estimating the accel noise to avoid increasing the
accel noise when the redisual is caused by an offset.


Former-commit-id: 0f1c7590e96fe1f2cf256f2c240f8cf55f726c59