Commander: make sure yaw airmode is not possible while arm gesture is enabled
It's unsafe to arm with the gesture when yaw airmode is enabled
because e.g. in Stabilized mode that results in a high yawrate setpoint
that the drone tries to follow even with zero thrust
because of the airmode.

It was handled before by checking the arm switch parameter because that
used to disable the stick arm gesture.
See d2fdd2567edaa1a51985f16c2678de060888b9f8 [formerly 24dc316973a049e50d3e6760eddbf7250cb55d2e]


Former-commit-id: 0153c870d2eb7539137992d0dd16d486210eb82f